app.py
import pickle
import pandas as pd
import streamlit as st


#load the data
df = pd.read_csv("cleaned_df.csv")


#load the pre trained model
with open("Log_reg_model.pkl","rb") as file:
    model = pickle.load(file)


#page setup
st.set_page_config(page_icon="💓",page_title="Heart Disease Predictor",layout="wide")

with st.sidebar:
    st.title("Heart Disease Predictor")
    #st.image("https://images.vexels.com/media/users/3/136170/isolated/preview/1a0fc726567fe21282676126358b795d-heart-disease-logo.png")
    st.image("https://www.bing.com/th/id/OGC.f037ee57d51ad51915799103937bc0b6?o=7&pid=1.7&rm=3&rurl=https%3a%2f%2fwww.gifcen.com%2fwp-content%2fuploads%2f2021%2f08%2fheart-gif.gif&ehk=Zw4Rq%2fpMtfisK0HlFF0vI5MBqB8pA9nHq%2fhEFdjqmPM%3d")


#user input

col1,col2 = st.columns(2)

with col1:
    age = st.number_input("Age: ",min_value=1,max_value=100)

    gender = st.radio("Gender: ",options=["Male","Female"],horizontal=True)
    gender = 1 if gender=="Male" else 0

    d = {"typical angina":0,"atypical angina":1,"non-anginal pain":2,"asymptomatic":3}
    chest_pain = st.selectbox("chest pain type: ",options=d)
    chest_pain = d[chest_pain] # to get numeric val

    resting_bp = st.number_input("Resting BP: ",min_value=80,max_value=250,step=5)
    cholestrol = st.number_input("Cholestrol: ",min_value=80,max_value=600,step=5)

    fbs = st.radio("Fasting Blood sugar: ",options=["Yes","No"],horizontal=True)
    fbs = 1 if fbs=="Yes" else 0

    d = {"Normal":0,"ST-T wave abnormality":1," left ventricular hypertrophy":2}
    restecg = st.selectbox("Rest ECG: ",options=d)
    restecg = d[restecg]


with col2:
    max = st.number_input("Max Heart rate: ",min_value=60,max_value=260,step=10)

    exang = st.radio("Exer induced angina: ",options=["Yes","No"],horizontal=True)
    exang = 1 if exang=="Yes" else 0

    oldpeak = st.number_input("Oldpeak: ",min_value=0.0,max_value=7.0,step=0.5)


    d = {"upsloping":0,"flat":1,"downsloping":2}
    slope = st.selectbox("Slope: ",options=d)
    slope = d[slope]

    num_major_vessels = st.selectbox("Num of major vessels: ",options=[0,1,2,3,4])


    d = {"normal":1,"fixed defect":2,"reversable defect":3}
    thal = st.selectbox("Thal: ",options=d)
    thal = d[thal]


c1,c2,c3 = st.columns([1.5,1,1])
if c2.button("Predict"):
    data = [[age,gender,chest_pain,resting_bp,cholestrol,fbs,restecg,max,exang,oldpeak,slope,num_major_vessels,thal]]
    pred = model.predict(data)[0]
    if pred==0:
        st.subheader("Low risk of Heart disease")
    else:
        st.subheader("High risk of Heart disease")
