requirements.txt
numpy
panda
seaborn
scikit-learn
streamlit
